various QOL changes

+++ utils/load_nrrd_dataset.py
@@ -64,7 +64,7 @@ class CTScanDataset(Dataset):
     @lru_cache(maxsize=1)
     def get_scan(self, scan_index: int) -> Tuple[np.array, dict]:
         data, metadata = nrrd.read(self.scans[scan_index])
-        return data.astype(np.float16), metadata
+        return data.astype(np.float32), metadata
 
     def __getitem__(self, index: int) -> Tuple[np.array, int]:
         data, metadata = self.get_scan(index)
